+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning][semver].
 
 ## [unreleased]
 
+## [4.1.8] - 2022-11-16
+
+### Added
+
+- [string] Add `normalizeName` && `normalizeCustom`: Removing non-alphanumeric chars from string
+- [string] Add `removeConsecutiveDuplicates`: Remove consecutive duplicates
+- [string] Add `trimAny`: Trim any characters
+
 ## [4.1.0] - 2022-11-10
 
 ### Changed